chore(migration): teams specs to ts
