--- tests/dashboard/teams/settings/teams-settings.spec.js
@@ -1,82 +1,0 @@
-const { mainTest } = require('../../../../fixtures');
-const { ProfilePage } = require('../../../../pages/profile-page.js');
-const { DashboardPage } = require('../../../../pages/dashboard/dashboard-page.js');
-const { TeamPage } = require('../../../../pages/dashboard/team-page.js');
-const { MainPage } = require('../../../../pages/workspace/main-page.js');
-const { qase } = require('playwright-qase-reporter/playwright');
-const { expect } = require('@playwright/test');
-const { createTeamName } = require('../../../../helpers/teams/create-team-name');
-
-const maxDiffPixelRatio = 0.001;
-
-let teamPage, dashboardPage, profilePage, mainPage;
-
-const team = createTeamName();
-
-mainTest.beforeEach(async ({ page }) => {
-  teamPage = new TeamPage(page);
-  dashboardPage = new DashboardPage(page);
-  profilePage = new ProfilePage(page);
-  mainPage = new MainPage(page);
-});
-
-mainTest(qase(1200, 'Team Settings - upload team profile picture'), async () => {
-  await teamPage.createTeam(team);
-  await teamPage.isTeamSelected(team);
-  await teamPage.openTeamSettingsPageViaOptionsMenu();
-  await teamPage.uploadTeamImage('images/images.png');
-  await teamPage.waitInfoMessageHidden();
-  await teamPage.hoverOnTeamName();
-  await expect(teamPage.teamIcon).toHaveScreenshot('team-profile-image.png', {
-    maxDiffPixelRatio: maxDiffPixelRatio,
-    mask: [teamPage.teamNameLabel],
-  });
-});
-
-mainTest(qase(1202, "Team. Settings - check 'Team members' info"), async () => {
-  await profilePage.openYourAccountPage();
-  await profilePage.isHeaderDisplayed('Your account');
-  await profilePage.changeProfileName('QA Engineer');
-  await profilePage.uploadProfileImage('images/sample.jpeg');
-  await profilePage.waitInfoMessageHidden();
-  await profilePage.backToDashboardFromAccount();
-  await teamPage.createTeam(team);
-  await teamPage.isTeamSelected(team);
-  await teamPage.openTeamSettingsPageViaOptionsMenu();
-
-  const teamOwner = 'QA Engineer (Owner)';
-  await teamPage.isTeamOwnerInfoDisplayed(teamOwner);
-  await teamPage.isTeamMembersInfoDisplayed('1 members');
-});
-
-mainTest(qase(1203, "Team. Settings - check 'Team projects' info"), async () => {
-  const projectFirst = 'QA Project 1';
-  const projectSecond = 'QA Project 2';
-
-  await teamPage.createTeam(team);
-  await teamPage.isTeamSelected(team);
-  await dashboardPage.createProject(projectFirst);
-  await dashboardPage.pinProjectByName(projectFirst);
-  await dashboardPage.createProject(projectSecond);
-  await dashboardPage.pinProjectByName(projectSecond);
-  await dashboardPage.openSidebarItem('Drafts');
-  await dashboardPage.createFileViaProjectPlaceholder();
-  await mainPage.backToDashboardFromFileEditor();
-  await dashboardPage.openPinnedProjectFromSidebar(projectFirst);
-  await dashboardPage.createFileViaProjectPlaceholder();
-  await mainPage.backToDashboardFromFileEditor();
-  await dashboardPage.openPinnedProjectFromSidebar(projectSecond);
-  await dashboardPage.createFileViaProjectPlaceholder();
-  await mainPage.backToDashboardFromFileEditor();
-
-  await teamPage.openTeamSettingsPageViaOptionsMenu();
-  await teamPage.isTeamProjectsInfoDisplayed('2 projects');
-  await teamPage.isTeamFilesInfoDisplayed('3 files');
-});
-
-mainTest(qase(1208, 'Delete a team via owner'), async () => {
-  await teamPage.createTeam(team);
-  await teamPage.isTeamSelected(team);
-  await teamPage.deleteTeam(team);
-  await teamPage.isTeamDeleted(team);
-});
