+++ tests/dashboard/teams/settings/teams-settings.spec.ts-snapshots/team-profile-image.png
