+++ tests/dashboard/teams/teams-viewer-role.spec.ts-snapshots/right-sidebar-image.png
